requirements.txt
requests
feedparser
python-dotenv
cryptography
beautifulsoup4
